src/main/java/com/venus/controllers/CalendarController.java
package com.venus.controllers;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.venus.feature.calendar.dto.CalendarResponse;
import com.venus.feature.calendar.service.CalendarService;

import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("calendar")
@RequiredArgsConstructor
public class CalendarController {

    private final CalendarService calendarService;

    @GetMapping
    public CalendarResponse getMyCalendar() {
        return calendarService.getMyCalendar();
    }
}
